# docs/specifications/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Dog Walking Management API
  description: |
    The Stardogwalker management API — Cardiff, South Wales.

    This API is designed **API-first** to support multiple interaction
    channels: web app, native mobile app, conversational chat interfaces
    (e.g. WhatsApp bot), and agentic AI applications. All business logic
    lives here; no channel-specific logic is embedded in the API.

    **Design principles:**
    - Stateless — every request authenticates via JWT bearer token
    - Fine-grained actions — discrete endpoints for each state transition
      (e.g. `/accept`, `/decline`) to support unambiguous intent from any
      client type including AI agents
    - Consistent response envelopes — list responses always include
      `data[]` and `pagination`; errors always return `code`, `message`,
      and optional `details`
    - Versioned — breaking changes will be introduced under `/v2` without
      removing `/v1` until all clients have migrated

    Covers: interest registration (public), authentication, dog owner and
    walker management, dog profiles, walk requests, walk scheduling and
    execution, real-time walk updates, and invoicing.
  version: 1.0.0
  contact:
    name: Dog Walking API Support
    email: support@dogwalking.example.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/v1
    description: Local development server
  - url: https://api.dogwalking.example.com/v1
    description: Production server
...
security:
  - bearerAuth: []
...
paths:
...
  /interest-requests:
    get:
      tags:
        - Interest Requests
      summary: List interest requests
      description: |
        Retrieve a paginated list of interest requests. Accessible by walkers only.
      operationId: listInterestRequests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: Filter interest requests by status
          schema:
            $ref: '#/components/schemas/InterestRequestStatus'
      responses:
        '200':
          description: A list of interest requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestRequestList'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Interest Requests
      summary: Submit an interest request
      description: |
        A prospective client submits their details via the public marketing
        site. No authentication is required.
      operationId: createInterestRequest
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterestRequestCreate'
      responses:
        '201':
          description: Interest request received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestRequestPublic'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /interest-requests/{interestRequestId}:
    parameters:
      - $ref: '#/components/parameters/InterestRequestId'
    get:
      tags:
        - Interest Requests
      summary: Get an interest request
      description: Retrieve details of a specific interest request.
      operationId: getInterestRequest
      responses:
        '200':
          description: Interest request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Interest Requests
      summary: Update an interest request
      description: |
        Allows the walker to add or update internal notes on an interest
        request before making a decision.
      operationId: updateInterestRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterestRequestUpdate'
      responses:
        '200':
          description: Interest request updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /interest-requests/{interestRequestId}/accept:
    parameters:
      - $ref: '#/components/parameters/InterestRequestId'
    post:
      tags:
        - Interest Requests
      summary: Accept an interest request
      description: |
        The walker accepts an interest request. The system automatically
        creates an Owner profile and User account from the request data and
        issues an invitation credential to the new client.
      operationId: acceptInterestRequest
      responses:
        '200':
          description: Interest request accepted; owner account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /interest-requests/{interestRequestId}/decline:
    parameters:
      - $ref: '#/components/parameters/InterestRequestId'
    post:
      tags:
        - Interest Requests
      summary: Decline an interest request
      description: The walker declines an interest request.
      operationId: declineInterestRequest
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterestRequestDecline'
      responses:
        '200':
          description: Interest request declined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterestRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Auth
  # ---------------------------------------------------------------------------
...
  /invoices:
    get:
      tags:
        - Invoices
      summary: List invoices
      description: |
        Retrieve a paginated list of invoices. Dog owners see their own invoices;
        walkers see invoices they have raised.
      operationId: listInvoices
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: Filter invoices by status
          schema:
            $ref: '#/components/schemas/InvoiceStatus'
        - name: ownerId
          in: query
          description: Filter invoices by dog owner
          schema:
            type: string
            format: uuid
        - name: walkerId
          in: query
          description: Filter invoices by dog walker
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A list of invoices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceList'
        '401':
          $ref: '#/components/responses/Unauthorised'

    post:
      tags:
        - Invoices
      summary: Create an invoice
      description: A dog walker raises an invoice for one or more completed walks.
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceRequest'
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
  /invoices/{invoiceId}:
    parameters:
      - $ref: '#/components/parameters/InvoiceId'
    get:
      tags:
        - Invoices
      summary: Get an invoice
      description: Retrieve details of a specific invoice.
      operationId: getInvoice
      responses:
        '200':
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Invoices
      summary: Update an invoice
      description: Update a draft invoice before it is sent.
      operationId: updateInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceUpdate'
      responses:
        '200':
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /invoices/{invoiceId}/send:
    parameters:
      - $ref: '#/components/parameters/InvoiceId'
    post:
      tags:
        - Invoices
      summary: Send an invoice
      description: Transition a draft invoice to `sent`, making it visible to the owner.
      operationId: sendInvoice
      responses:
        '200':
          description: Invoice sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /invoices/{invoiceId}/pay:
    parameters:
      - $ref: '#/components/parameters/InvoiceId'
    post:
      tags:
        - Invoices
      summary: Mark an invoice as paid
      description: Record payment against an invoice.
      operationId: payInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoicePayment'
      responses:
        '200':
          description: Invoice marked as paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

# =============================================================================
# Components
# =============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    InterestRequestId:
      name: interestRequestId
      in: path
      required: true
      description: Unique identifier of the interest request
      schema:
        type: string
        format: uuid
...
    InvoiceId:
      name: invoiceId
      in: path
      required: true
      description: Unique identifier of the invoice
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Number of results per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validationError:
              summary: Validation error
              value:
                code: VALIDATION_ERROR
                message: One or more fields failed validation.
                details:
                  - field: email
                    issue: Must be a valid email address.
                  - field: durationMinutes
                    issue: Must be between 15 and 240.
            noMatchingRate:
              summary: No matching rate found
              value:
                code: NO_MATCHING_RATE
                message: No rate found for this walk type, dog count, and duration. Please update the walker's rate card.
            invalidStateTransition:
              summary: Invalid state transition
              value:
                code: INVALID_STATE_TRANSITION
                message: This action cannot be performed on a resource in its current status.
            schedulingConflict:
              summary: Scheduling conflict
              value:
                code: SCHEDULING_CONFLICT
                message: The walker already has an overlapping walk scheduled for this time.
    Unauthorised:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingToken:
              summary: Missing bearer token
              value:
                code: AUTHENTICATION_REQUIRED
                message: A valid bearer token is required to access this resource.
            invalidCredentials:
              summary: Invalid login credentials
              value:
                code: INVALID_CREDENTIALS
                message: The email or password provided is incorrect.
            tokenExpired:
              summary: Expired token
              value:
                code: TOKEN_EXPIRED
                message: The access token has expired. Please refresh your session.
    Forbidden:
      description: Authenticated user does not have permission to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            insufficientRole:
              summary: Insufficient role
              value:
                code: FORBIDDEN
                message: Your role does not have permission to perform this action.
            notOwnResource:
              summary: Not own resource
              value:
                code: FORBIDDEN
                message: You can only access your own records.
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            resourceNotFound:
              summary: Resource not found
              value:
                code: RESOURCE_NOT_FOUND
                message: The requested resource could not be found.
    Conflict:
      description: The request conflicts with the current state of the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            duplicateEmail:
              summary: Duplicate email
              value:
                code: DUPLICATE_EMAIL
                message: A user with this email address already exists.
            alreadyAccepted:
              summary: Already accepted
              value:
                code: ALREADY_PROCESSED
                message: This request has already been accepted or declined.

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    InterestRequestStatus:
      type: string
      enum:
        - pending
        - accepted
        - declined
      example: pending
    InterestRequestCreate:
      type: object
      required:
        - firstName
        - lastName
        - email
        - phone
        - postcode
        - dogDescription
      properties:
        firstName:
          type: string
          example: Sarah
        lastName:
          type: string
          example: Davies
        email:
          type: string
          format: email
          example: sarah.davies@example.com
        phone:
          type: string
          example: +44 7700 900789
        postcode:
          type: string
          description: Must be within the Cardiff / South Wales service area
          example: CF10 1AB
        dogDescription:
          type: string
          description: Brief description of the prospective client's dog(s)
          example: I have a 2-year-old Golden Retriever called Buddy. Very friendly, no medical issues.
    InterestRequestPublic:
      type: object
      description: Confirmation returned to the prospective client after submission
      required:
        - id
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: aabb1122-ccdd-3344-eeff-556677889900
        createdAt:
          type: string
          format: date-time
          example: '2026-03-28T10:00:00Z'
    InterestRequestUpdate:
      type: object
      properties:
        notes:
          type: string
          description: Internal walker notes (not visible to the prospective client)
          example: Spoke to Sarah on the phone. Buddy sounds like a great fit.
    InterestRequestDecline:
      type: object
      properties:
        reason:
          type: string
          description: Reason for declining the interest request
          example: We are not currently covering the CF5 area.
    InterestRequest:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - phone
        - postcode
        - dogDescription
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: aabb1122-ccdd-3344-eeff-556677889900
        firstName:
          type: string
          example: Sarah
        lastName:
          type: string
          example: Davies
        email:
          type: string
          format: email
          example: sarah.davies@example.com
        phone:
          type: string
          example: +44 7700 900789
        postcode:
          type: string
          example: CF10 1AB
        dogDescription:
          type: string
          example: I have a 2-year-old Golden Retriever called Buddy.
        status:
          $ref: '#/components/schemas/InterestRequestStatus'
        declineReason:
          type: string
          nullable: true
          example: null
        notes:
          type: string
          nullable: true
          description: Internal walker notes; not returned to unauthenticated callers
          example: Spoke to Sarah on the phone.
        ownerId:
          type: string
          format: uuid
          nullable: true
          description: Set when the interest request is accepted and an owner account is created
          example: null
        createdAt:
          type: string
          format: date-time
          example: '2026-03-28T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-03-28T10:00:00Z'
    InterestRequestList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/InterestRequest'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # --- Shared ---------------------------------------------------------------
...
    Pagination:
      type: object
      required:
        - page
        - pageSize
        - total
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          description: Total number of matching records
          example: 42
...
    InvoiceStatus:
      type: string
      enum:
        - draft
        - sent
        - paid
        - overdue
        - cancelled
      example: sent
...
    InvoiceUpdate:
      type: object
      properties:
        ownerId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        walkerId:
          type: string
          format: uuid
          example: 7b1e4d2a-1234-5678-abcd-ef0123456789
        lineItems:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
        dueDate:
          type: string
          format: date
          description: Payment due date
          example: '2025-04-30'
        notes:
          type: string
          description: Additional notes on the invoice
          example: Thank you for choosing Happy Paws Walking Service!
    InvoiceRequest:
      type: object
      required:
        - ownerId
        - walkerId
        - lineItems
        - dueDate
      properties:
        ownerId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        walkerId:
          type: string
          format: uuid
          example: 7b1e4d2a-1234-5678-abcd-ef0123456789
        lineItems:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
        dueDate:
          type: string
          format: date
          description: Payment due date
          example: '2025-04-30'
        notes:
          type: string
          description: Additional notes on the invoice
          example: Thank you for choosing Happy Paws Walking Service!
...
    Invoice:
      type: object
      required:
        - id
        - ownerId
        - walkerId
        - lineItems
        - subtotal
        - taxRate
        - taxAmount
        - total
        - dueDate
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: ffee1122-3344-5566-7788-99aabbccddee
        invoiceNumber:
          type: string
          description: Human-readable invoice number
          example: INV-2025-0042
        ownerId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        walkerId:
          type: string
          format: uuid
          example: 7b1e4d2a-1234-5678-abcd-ef0123456789
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineItem'
        subtotal:
          type: number
          format: float
          description: Sum of all line item totals before any tax
          example: 45.00
        taxRate:
          type: number
          format: float
          description: Tax rate applied as a percentage
          example: 0.00
        taxAmount:
          type: number
          format: float
          example: 0.00
        total:
          type: number
          format: float
          description: Total amount due including tax
          example: 45.00
        dueDate:
          type: string
          format: date
          example: '2025-04-30'
        status:
          $ref: '#/components/schemas/InvoiceStatus'
        notes:
          type: string
          nullable: true
          example: Thank you for choosing Happy Paws Walking Service!
        paidAt:
          type: string
          format: date-time
          nullable: true
          example: null
        paymentMethod:
          type: string
          nullable: true
          example: null
        paymentReference:
          type: string
          nullable: true
          example: null
        createdAt:
          type: string
          format: date-time
          example: '2025-04-16T09:00:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-04-16T09:00:00Z'
    InvoiceList:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Invoice'
        pagination:
          $ref: '#/components/schemas/Pagination'